Refactor react-native-client-sdk
