+++ packages/react-native-client/src/common/client.ts
@@ -1,18 +1,14 @@
 import { Metadata } from '../RNFishjamClient.types';
 import RNFishjamClientModule from '../RNFishjamClientModule';
 
-export const connect = (url: string, peerToken: string) => {
-  return RNFishjamClientModule.connect(url, peerToken);
-};
-
-export const joinRoom = (peerMetadata: Metadata) => {
-  return RNFishjamClientModule.joinRoom(peerMetadata);
+export const connect = (
+  url: string,
+  peerToken: string,
+  peerMetadata: Metadata,
+) => {
+  return RNFishjamClientModule.connect(url, peerToken, peerMetadata);
 };
 
 export const leaveRoom = () => {
   return RNFishjamClientModule.leaveRoom();
-};
-
-export const cleanUp = () => {
-  return RNFishjamClientModule.cleanUp();
 };
