+++ packages/react-native-client/src/common/webRTC.ts
@@ -1,4 +1,4 @@
-import { LoggingSeverity, TrackEncoding } from '../RNFishjamClient.types';
+import { TrackEncoding } from '../RNFishjamClient.types';
 import RNFishjamClientModule from '../RNFishjamClientModule';
 
 /**
@@ -15,15 +15,4 @@ export async function setTargetTrackEncoding(
   encoding: TrackEncoding,
 ) {
   await RNFishjamClientModule.setTargetTrackEncoding(trackId, encoding);
-}
-
-/**
- * Function that changes level of debugging logs in WebRTC.
- * @param severity to use when displaying logs
- * @returns a promise that is resolved when debug severity is changed
- */
-export function changeWebRTCLoggingSeverity(
-  severity: LoggingSeverity,
-): Promise<void> {
-  return RNFishjamClientModule.changeWebRTCLoggingSeverity(severity);
 }
