+++ packages/react-native-client/tsconfig.json
@@ -4,6 +4,6 @@
   "compilerOptions": {
     "outDir": "./build"
   },
-  "include": ["./src"],
+  "include": ["./src", "./debug"],
   "exclude": ["**/__mocks__/*", "**/__tests__/*"]
 }
